+++ php/control_clientes.php
@@ -131,15 +131,24 @@ switch ($Accion) {
     	$User = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM `users` WHERE user_id = $id_user"));
     	//SE VERIFICA SI EL USUARIO LOGEADO TIENE PERMISO DE BORRAR CLIENTES
     	if ($User['b_clientes'] == 1) {
-    		#VERIFICAMOS QUE SE BORRE CORRECTAMENTE EL CLIENTE DE `punto-venta_clientes`
-			if(mysqli_query($conn, "DELETE FROM `punto-venta_clientes` WHERE `punto-venta_clientes`.`id` = $id")){
-			  #SI ES ELIMINADO MANDAR MSJ CON ALERTA
-			  echo '<script >M.toast({html:"Cliente borrado con exito.", classes: "rounded"})</script>';
-			}else{
-			  #SI NO ES BORRADO MANDAR UN MSJ CON ALERTA
-			  echo "<script >M.toast({html: 'Ha ocurrido un error.', classes: 'rounded'});/script>";
+    		#SELECCIONAMOS LA INFORMACION A BORRAR
+    		$cliente = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM `punto-venta_clientes` WHERE id = $id"));
+    		#CREAMOS EL SQL DE LA INSERCION A LA TABLA  `pv_borrar_cliente` PARA NO PERDER INFORMACION
+			$sql = "INSERT INTO `pv_borrar_cliente` (nombre, telefono, direccion, colonia, cp, rfc, email, localidad, registro, borro, fecha_borro) 
+				VALUES('".$cliente['nombre']."', '".$cliente['telefono']."', '".$cliente['direccion']."', '".$cliente['colonia']."', '".$cliente['cp']."', '".$cliente['rfc']."', '".$cliente['email']."', '".$cliente['localidad']."', '".$cliente['usuario']."', '$id_user','$Fecha_hoy')";
+			//VERIFICAMOS QUE LA SENTECIA FUE EJECUTADA CON EXITO!
+			if(mysqli_query($conn, $sql)){
+				//SI DE CREA LA INSERCION PROCEDEMOS A BORRRAR DE LA TABLA `punto-venta_clientes`
+	    		#VERIFICAMOS QUE SE BORRE CORRECTAMENTE EL CLIENTE DE `punto-venta_clientes`
+				if(mysqli_query($conn, "DELETE FROM `punto-venta_clientes` WHERE `punto-venta_clientes`.`id` = $id")){
+				  #SI ES ELIMINADO MANDAR MSJ CON ALERTA
+				  echo '<script >M.toast({html:"Cliente borrado con exito.", classes: "rounded"})</script>';
+				}else{
+				  #SI NO ES BORRADO MANDAR UN MSJ CON ALERTA
+				  echo "<script >M.toast({html: 'Ha ocurrido un error.', classes: 'rounded'});/script>";
+				}
+				echo '<script>recargar_clientes()</script>';// REDIRECCIONAMOS (FUNCION ESTA EN ARCHIVO modals.php)
 			}
-			echo '<script>recargar_clientes()</script>';// REDIRECCIONAMOS (FUNCION ESTA EN ARCHIVO modals.php)
 	    }else{
 			echo '<script >M.toast({html:"Permiso denegado.", classes: "rounded"});
 			M.toast({html:"Comunicate con un administrador.", classes: "rounded"});</script>';
